fix: geofence Range int -> double 로 수정

+++ src/main/java/com/kbe5/rento/domain/device/dto/resonse/GeofenceControlInfoResponse.java
@@ -10,7 +10,7 @@ public record GeofenceControlInfoResponse(
         int updateValue,
         int geofenceGroupId,
         short geoEventType,
-        int geofenceRange,
+        double geofenceRange,
         BigDecimal latitude,
         BigDecimal longitude,
         LocalDateTime onTime,
